refactor: rename user test classes and update package structure for better organization
