+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/SessionManagementTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.admin;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
