+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/MinistryViewTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
 
 import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
