+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/PartyViewTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
