+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/SearchViewTest.java
@@ -1,11 +1,10 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
 
-
-import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
