+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/UserAuthenticationTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
 
 import java.util.UUID;
